wip - get server running, instructions for installing ssl cert
