+++ zap/install_linux.go
@@ -1,0 +1,9 @@
+package zap
+
+func installService(httpAddr, httpsAddr string) error {
+	return nil
+}
+
+func uninstallService() error {
+	return nil
+}
